+++ KriptoAlgoritmi/SimpleSubcpp.cpp
@@ -1,20 +1,20 @@
 #include "SimpleSub.h"
 
-SimpleSub::SimpleSub(unsigned short Lenght, uint16_t* String, uint16_t* Map)
+SimpleSub::SimpleSub(unsigned short Lenght, uint8_t* String, uint8_t* Map)
 {
-	this->charString = (uint16_t*)malloc(Lenght * sizeof(uint16_t));
-	this->charMap = (uint16_t*)malloc(Lenght * sizeof(uint16_t));
+	this->charString = (uint8_t*)malloc(Lenght * sizeof(uint8_t));
+	this->charMap = (uint8_t*)malloc(Lenght * sizeof(uint8_t));
 
-	memcpy(this->charString, String, Lenght * sizeof(uint16_t));
-	memcpy(this->charMap, Map, Lenght * sizeof(uint16_t));
+	memcpy(this->charString, String, Lenght * sizeof(uint8_t));
+	memcpy(this->charMap, Map, Lenght * sizeof(uint8_t));
 
 	this->lenght = Lenght;
 }
 
-uint16_t* SimpleSub::Encode(uint16_t* string, unsigned long lenght)
+uint8_t* SimpleSub::Encode(uint8_t* string, unsigned long lenght)
 {
-	uint16_t* data = (uint16_t*)malloc(lenght * sizeof(uint16_t));
-	memcpy(data, string, lenght * sizeof(uint16_t));
+	uint8_t* data = (uint8_t*)malloc(lenght * sizeof(uint8_t));
+	memcpy(data, string, lenght * sizeof(uint8_t));
 
 	for (int i = 0; i < lenght; i++)
 	{
@@ -24,10 +24,10 @@ uint16_t* SimpleSub::Encode(uint16_t* string, unsigned long lenght)
 	return data;
 }
 
-uint16_t* SimpleSub::Decode(uint16_t* string, unsigned long lenght)
+uint8_t* SimpleSub::Decode(uint8_t* string, unsigned long lenght)
 {
-	uint16_t* data = (uint16_t*)malloc(lenght * sizeof(uint16_t));
-	memcpy(data, string, lenght * sizeof(uint16_t));
+	uint8_t* data = (uint8_t*)malloc(lenght * sizeof(uint8_t));
+	memcpy(data, string, lenght * sizeof(uint8_t));
 
 	for (int i = 0; i < lenght; i++)
 	{
@@ -37,7 +37,7 @@ uint16_t* SimpleSub::Decode(uint16_t* string, unsigned long lenght)
 	return data;
 }
 
-uint16_t SimpleSub::find(uint16_t data)
+uint8_t SimpleSub::find(uint8_t data)
 {
 	for (int i = 0; i < this->lenght; i++)
 	{
@@ -50,7 +50,7 @@ uint16_t SimpleSub::find(uint16_t data)
 	return data;
 }
 
-uint16_t SimpleSub::findD(uint16_t data)
+uint8_t SimpleSub::findD(uint8_t data)
 {
 	for (int i = 0; i < this->lenght; i++)
 	{
